requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cloud-init==22.3.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
discord==2.1.0
discord.py==2.1.0
distro==1.7.0
distro-info===1.1build1
Django==3.2.9
docopt==0.6.2
Flag==0.1.1
flake8==6.0.0
fonttools==4.38.0
frozenlist==1.3.3
giphy-client==1.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
matplotlib==3.6.2
mccabe==0.7.0
more-itertools==8.10.0
MouseInfo==0.1.3
multidict==6.0.3
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==22.0
pandas==1.5.2
pexpect==4.8.0
Pillow==8.4.0
pipreqs==0.4.11
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycountry==22.3.5
pyflakes==3.0.1
pygame==2.1.0
PyGetWindow==0.0.9
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.28
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
python-utils==3.4.5
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.4
sqlparse==0.4.2
ssh-import-id==5.11
systemd-python==234
tabulate==0.9.0
Twisted==22.1.0
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.8
wadllib==1.3.6
wordcloud==1.8.2.2
yapf==0.32.0
yarg==0.1.9
yarl==1.8.2
zipp==1.0.0
zope.interface==5.4.0
